libhwc2.1: Correct the event type assignment
The 4 MSB will be used to denote the group, and the remaining 28 bits
will be used to denote the group member.

For this requirement, we may need to drop a group of events using a
group mask. For example, with a group mask of 0x10000000, the first
event is 0x10000001, the second event is 0x10000002, and the third event
is 0x10000004. Using drop(group mask) will drop all these events.

Bug: 351154317
Test: check the correctness of drop a single event or a group of events
Change-Id: I25c7dafad73c72e4c04200bff4ceb107e5d7bdf1