libhwc2.1: maintain the consistency of present timeout controller
|mPresentTimeoutController| was previously used in debug commands to
differentiate  between DDIC and software as the present timeout
controller.

In special cases, like when the minimum refresh rate is set, we may
configure DDIC to  auto mode as the present timeout controller.
We only need to unset auto mode when exiting the minimum refresh rate
and when |mDefaultPresentTimeoutController| is set to |software|;
otherwise, maintain auto mode.

This change ensures consistent behavior of the present timeout
controller  when the minimum refresh rate is set.

Bug: 357165904
Test: verify |mPresentTimeoutController| is consistent to auto/manual
mode

Change-Id: Ib19072e5379ce3ddc975731aed6e89ef0e4df319