libhwc2.1: categorize 1Hz refresh correctly
Since we have introduced non-present refresh, when calculating 1 Hz
refresh, it will automatically be categorized as the last refresh type.
If the last refresh is a non-present refresh, this will lead to
incorrect statistics.
In this case, when calculating a 1 Hz refresh, we set the refresh type
to present to resolve the issue.

Bug: 353976206
Test: ensure that a 1 Hz refresh rate is not mistakenly categorized as a
non-present refresh.

Change-Id: Id335f76235bb433bce1347761ec743fef60c7ee4