libhwc2.1: recursively manage vendor present timeout.
Previously, after the first present timeout occurred, we scheduled all
handling events one at a time.
However, this approach is inefficient, as many events would be dropped
if a consecutive present occurred.
In this implementation, we recursively schedule the handling events. We
post one event at a time and recursively post the following handling
event if necessary.
This approach reduces the number of events posted and dropped, thereby
decreasing system load.
Additionally, it is better with the scheduler we plan to implement later
.

Bug: 352833950
Test: check that the FI patterns are the same as before and that only
one event is scheduled instead of multiple events
Change-Id: I758eece338461944b2048aa4e5ba6213b8e62f79

Change-Id: I4ae1b405395a1d9799dbac534ef9f4854f926ec4