hwc: libvrr: Set minimum refresh rate only with correct configuration
The minimum refresh rate should only be set once active configuration
is correctly applied. If not, cadence alignment issues may occur,
potentially leading to jank issue.
If the TE frequency of the current configuration does not meet the
required criteria, the minimum refresh rate request will be temporarily
suspended. A timer will be set, and we expect the target configuration
to activate before the timeout.
If a timeout occurs, an error message will be displayed, and the minimum
refresh rate request will be dropped.

Bug: 376330991
Test: ensure no minimum refresh rate is only set after the correct
configuration is applied.

Change-Id: I1852139039d1b55a206d1c65858e1adee7949c86